requirements.txt
mysql-connector-python
flask
flask-session